+++ massiv-test/tests/Test/Massiv/VectorSpec.hs
@@ -46,23 +46,23 @@ withSeed (SeedVector seed) f = runST $ do
   gen <- MWC.initialize seed
   f gen
 
-prop_replicateM :: SeedVector -> Int -> Property
-prop_replicateM seed k =
-  withSeed @(V.Vector DS Word) seed (V.replicateM (Sz k) . uniform)
+prop_sreplicateM :: SeedVector -> Int -> Property
+prop_sreplicateM seed k =
+  withSeed @(V.Vector DS Word) seed (V.sreplicateM (Sz k) . uniform)
   !==! withSeed seed (VP.replicateM k . uniform)
 
-prop_generateM :: SeedVector -> Int -> Fun Int Word -> Property
-prop_generateM seed k f =
-  withSeed @(V.Vector DS Word) seed (genWith (V.generateM (Sz k)))
+prop_sgenerateM :: SeedVector -> Int -> Fun Int Word -> Property
+prop_sgenerateM seed k f =
+  withSeed @(V.Vector DS Word) seed (genWith (V.sgenerateM (Sz k)))
   !==! withSeed seed (genWith (VP.generateM k))
   where
     genWith :: PrimMonad f => ((Int -> f Word) -> t) -> MWC.Gen (PrimState f) -> t
     genWith genM gen = genM (\i -> xor (apply f i) <$> uniform gen)
 
 
-prop_iterateNM :: SeedVector -> Int -> Word -> Property
-prop_iterateNM seed k a =
-  withSeed @(V.Vector DS Word) seed (genWith (\action -> V.iterateNM (Sz k) action a))
+prop_siterateNM :: SeedVector -> Int -> Word -> Property
+prop_siterateNM seed k a =
+  withSeed @(V.Vector DS Word) seed (genWith (\action -> V.siterateNM (Sz k) action a))
   !==! withSeed seed (genWith (\action -> VP.iterateNM k action a))
   where
     genWith :: PrimMonad f => ((Word -> f Word) -> t) -> MWC.Gen (PrimState f) -> t
@@ -83,33 +83,41 @@ spec = do
             V.tail' arr !!==!! VP.tail (toPrimitiveVector arr)
           prop "take" $ \n (arr :: Array P Ix1 Word) ->
             V.take (Sz n) arr !==! VP.take n (toPrimitiveVector arr)
+          prop "stake" $ \n (arr :: Array P Ix1 Word) ->
+            V.stake (Sz n) arr !==! VP.take n (toPrimitiveVector arr)
           prop "drop" $ \n (arr :: Array P Ix1 Word) ->
             V.drop (Sz n) arr !==! VP.drop n (toPrimitiveVector arr)
-          prop "splitAt" $ \sz (arr :: Array P Ix1 Word) ->
-            let (larr, rarr) = V.splitAt (Sz sz) arr
+          prop "sdrop" $ \n (arr :: Array P Ix1 Word) ->
+            V.sdrop (Sz n) arr !==! VP.drop n (toPrimitiveVector arr)
+          prop "sliceAt" $ \sz (arr :: Array P Ix1 Word) ->
+            let (larr, rarr) = V.sliceAt (Sz sz) arr
                 (lvec, rvec) = VP.splitAt sz (toPrimitiveVector arr)
              in (larr !==! lvec) .&&. (rarr !==! rvec)
       describe "Constructors" $ do
         describe "Initialization" $ do
           it "empty" $ toPrimitiveVector (V.empty :: V.Vector P Word) `shouldBe` VP.empty
           prop "singleton" $ \e -> (V.singleton e :: V.Vector P Word) !==! VP.singleton e
+          prop "ssingleton" $ \(e :: Word) -> V.ssingleton e !==! VP.singleton e
           prop "replicate" $ \comp k (e :: Word) -> V.replicate comp (Sz k) e !==! VP.replicate k e
+          prop "sreplicate" $ \k (e :: Word) -> V.sreplicate (Sz k) e !==! VP.replicate k e
           prop "generate" $ \comp k (f :: Fun Int Word) ->
             V.generate comp (Sz k) (apply f) !==! VP.generate k (apply f)
-          prop "iterateN" $ \n (f :: Fun Word Word) a ->
-            V.iterateN (Sz n) (apply f) a !==! VP.iterateN n (apply f) a
+          prop "sgenerate" $ \k (f :: Fun Int Word) ->
+            V.sgenerate (Sz k) (apply f) !==! VP.generate k (apply f)
+          prop "siterateN" $ \n (f :: Fun Word Word) a ->
+            V.siterateN (Sz n) (apply f) a !==! VP.iterateN n (apply f) a
         describe "Monadic initialization" $ do
-          prop "replicateM" prop_replicateM
-          prop "generateM" prop_generateM
-          prop "iterateNM" prop_iterateNM
+          prop "sreplicateM" prop_sreplicateM
+          prop "sgenerateM" prop_sgenerateM
+          prop "siterateNM" prop_siterateNM
         describe "Unfolding" $ do
-          prop "unfoldr" $ \(a :: Word) ->
+          prop "sunfoldr" $ \(a :: Word) ->
             let f b
                   | b > 10000 || b `div` 17 == 0 = Nothing
                   | otherwise = Just (b * b, b + 1)
-             in V.unfoldr f a !==! VP.unfoldr f a
-          prop "unfoldrN" $ \n (a :: Word) ->
+             in V.sunfoldr f a !==! VP.unfoldr f a
+          prop "sunfoldrN" $ \n (a :: Word) ->
             let f b
                   | b > 10000 || b `div` 19 == 0 = Nothing
                   | otherwise = Just (b * b, b + 1)
-             in V.unfoldrN (Sz n) f a !==! VP.unfoldrN n f a
+             in V.sunfoldrN (Sz n) f a !==! VP.unfoldrN n f a